Fix bug 67630 "kaudiocreator doesn't read cddb cache"
The problem was that even though libkcddb looked in the cache, it looked
for the wrong discid, because the length was wrong.
The reason that it worked for remote lookup was that the rest of the numbers
were correct, so the server could still find possible matches.

CCMAIL: 67630-done@bugs.kde.org

svn path=/trunk/kdemultimedia/kaudiocreator/; revision=266173